#rdeploy: wip on dev api
