+++ api/infra/config.js
@@ -1,10 +1,10 @@
 const config = {
-    repositoryName: "binaryx_app",
-    containerName: "binaryx_app",
+    repositoryName: "binaryx_api",
+    containerName: "binaryx_api",
     profile: "binaryx",
     region: "eu-west-1",
     accountId: "292041122161.dkr.ecr.eu-west-1.amazonaws.com",
-    dockerFile: 'Dockerfile.prod'
+    dockerFile: 'Dockerfile'
 }
 
 module.exports = {
