+++ api/src/app/users/user.entity.ts
@@ -5,37 +5,28 @@ import {
   CreateDateColumn,
   UpdateDateColumn,
 } from 'typeorm';
-import { ObjectType, Field } from '@nestjs/graphql';
 
-@ObjectType()
 @Entity()
 export class User {
-  @Field()
   @PrimaryGeneratedColumn()
   id: number;
 
-  @Field()
   @Column({ nullable: false })
   provider: string;
 
-  @Field()
   @Column({ nullable: false })
   providerId: string;
 
-  @Field()
   @Column({ nullable: false })
   username: string;
 
-  @Field()
   @Column({ nullable: false })
   name?: string;
 
-  @Field()
   @Column()
   @CreateDateColumn()
   created_at: Date;
 
-  @Field()
   @Column()
   @UpdateDateColumn()
   updated_at: Date;
