--- infra/scripts/autoReleaseDocker.js
@@ -1,6 +1,0 @@
-#!/usr/bin/env node
-
-const {config} = require("../src/config");
-const {appAutoReleaseDocker} = require("../src/docker");
-
-appAutoReleaseDocker(config)
